+++ android/src/ti/admob/AdmobModule.java
@@ -8,46 +8,42 @@
 
 package ti.admob;
 
-import android.os.AsyncTask;
-
-import java.io.IOException;
-
-import com.google.android.gms.ads.identifier.AdvertisingIdClient;
-import com.google.android.gms.common.GooglePlayServicesUtil;
-import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
-import com.google.android.gms.common.GooglePlayServicesRepairableException;
-
-import org.appcelerator.kroll.KrollDict;
-import org.appcelerator.kroll.KrollFunction;
-import org.appcelerator.kroll.KrollModule;
-import org.appcelerator.kroll.annotations.Kroll;
-import org.appcelerator.kroll.common.Log;
-import org.appcelerator.titanium.TiApplication;
-
-import org.appcelerator.kroll.KrollFunction;
-import org.appcelerator.kroll.KrollDict;
-import org.appcelerator.kroll.KrollObject;
-import org.appcelerator.titanium.TiApplication;
-
-import java.util.List;
-import java.util.Arrays;
-import java.util.ArrayList;
-import java.net.URL;
-import java.net.MalformedURLException;
-
-import android.net.Uri;
-import android.content.Context;
 import android.app.Activity;
-
-import com.google.android.gms.common.GooglePlayServicesUtil;
-
-import com.google.ads.consent.ConsentInformation;
+import android.content.Context;
+import android.net.Uri;
+import android.os.AsyncTask;
+import android.os.Bundle;
+import com.google.ads.consent.AdProvider;
 import com.google.ads.consent.ConsentForm;
 import com.google.ads.consent.ConsentFormListener;
 import com.google.ads.consent.ConsentInfoUpdateListener;
+import com.google.ads.consent.ConsentInformation;
 import com.google.ads.consent.ConsentStatus;
 import com.google.ads.consent.DebugGeography;
-import com.google.ads.consent.AdProvider;
+import com.google.ads.mediation.admob.AdMobAdapter;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.MobileAds;
+import com.google.android.gms.ads.identifier.AdvertisingIdClient;
+import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
+import com.google.android.gms.common.GooglePlayServicesRepairableException;
+import com.google.android.gms.common.GooglePlayServicesUtil;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import org.appcelerator.kroll.KrollDict;
+import org.appcelerator.kroll.KrollFunction;
+import org.appcelerator.kroll.KrollModule;
+import org.appcelerator.kroll.KrollObject;
+import org.appcelerator.kroll.annotations.Kroll;
+import org.appcelerator.kroll.common.Log;
+import org.appcelerator.titanium.TiApplication;
+import org.appcelerator.titanium.TiBlob;
+import org.appcelerator.titanium.io.TiBaseFile;
+import org.appcelerator.titanium.util.TiConvert;
 
 @Kroll.module(name = "Admob", id = "ti.admob")
 public class AdmobModule extends KrollModule
@@ -65,16 +61,24 @@ public class AdmobModule extends KrollModule
 
 	public static String PUBLISHER_ID;
 
-	// *
+	// properties
+	public static String PROPERTY_AD_SIZE = "adSize";
+	public static String PROPERTY_AD_UNIT_ID = "adUnitId";
+	public static String PROPERTY_DEBUG_ENABLED = "debugEnabled";
 	public static String PROPERTY_COLOR_BG = "adBackgroundColor";
 	public static String PROPERTY_COLOR_BG_TOP = "backgroundTopColor";
 	public static String PROPERTY_COLOR_BORDER = "borderColor";
 	public static String PROPERTY_COLOR_TEXT = "textColor";
 	public static String PROPERTY_COLOR_LINK = "linkColor";
 	public static String PROPERTY_COLOR_URL = "urlColor";
+	public static String PROPERTY_IS_TAGGED_FOR_UNDER_AGE_OF_CONSENT = "isTaggedForUnderAgeOfConsent";
 
-	public static String PROPERTY_COLOR_TEXT_DEPRECATED = "primaryTextColor";
-	public static String PROPERTY_COLOR_LINK_DEPRECATED = "secondaryTextColor";
+	// new event constans
+	public static final String EVENT_AD_LOAD = "load";
+	public static final String EVENT_AD_FAIL = "fail";
+	public static final String EVENT_AD_CLOSED = "close";
+	public static final String EVENT_AD_OPENED = "open";
+	public static final String EVENT_AD_LEFT_APP = "leftapp";
 
 	public AdmobModule()
 	{
@@ -87,6 +91,12 @@ public class AdmobModule extends KrollModule
 	public static final String AD_RECEIVED = "ad_received";
 	@Kroll.constant
 	public static final String AD_NOT_RECEIVED = "ad_not_received";
+	@Kroll.constant
+	public static final String AD_CLOSED = "ad_closed";
+	@Kroll.constant
+	public static final String AD_OPENED = "ad_opened";
+	@Kroll.constant
+	public static final String AD_LEFT_APPLICATION = "ad_left_application";
 
 	// Response from "isGooglePlayServicesAvailable()""
 	@Kroll.constant
@@ -108,6 +118,7 @@ public class AdmobModule extends KrollModule
 	@Kroll.constant
 	public static final int CONSENT_STATUS_PERSONALIZED = 2;
 
+	// Debug geography constants
 	@Kroll.constant
 	public static final int DEBUG_GEOGRAPHY_DISABLED = 0;
 	@Kroll.constant
@@ -115,9 +126,39 @@ public class AdmobModule extends KrollModule
 	@Kroll.constant
 	public static final int DEBUG_GEOGRAPHY_NOT_EEA = 3;
 
+	// AdSize constants
+	@Kroll.constant
+	public static final int AD_SIZE_BANNER = 0;
+	@Kroll.constant
+	public static final int AD_SIZE_FLUID = 1;
+	@Kroll.constant
+	public static final int AD_SIZE_FULL_BANNER = 2;
+	@Kroll.constant
+	public static final int AD_SIZE_LARGE_BANNER = 3;
+	@Kroll.constant
+	public static final int AD_SIZE_LEADERBOARD = 4;
+	@Kroll.constant
+	public static final int AD_SIZE_MEDIUM_RECTANGLE = 5;
+	@Kroll.constant
+	public static final int AD_SIZE_SEARCH = 6;
+	@Kroll.constant
+	public static final int AD_SIZE_SMART_BANNER = 7;
+	@Kroll.constant
+	public static final int AD_SIZE_WIDE_SKYSCRAPER = 8;
+
+	@Kroll.constant
+	public static final String SIMULATOR_ID = AdRequest.DEVICE_ID_EMULATOR;
+
+	@Kroll.method
+	public void initialize(String appID)
+	{
+		MobileAds.initialize(TiApplication.getInstance(), appID);
+	}
+
 	@Kroll.method
 	public int isGooglePlayServicesAvailable()
 	{
+		Log.w(TAG, "isGooglePlayServices in ti.admob is deprecated. Use the same method from ti.playservices instead.");
 		return GooglePlayServicesUtil.isGooglePlayServicesAvailable(TiApplication.getAppRootOrCurrentActivity());
 	}
 
@@ -139,20 +180,24 @@ public class AdmobModule extends KrollModule
 	}
 
 	@Kroll.method
-	public void isLimitAdTrackingEnabled(KrollFunction callback) {
+	public void isLimitAdTrackingEnabled(KrollFunction callback)
+	{
 		if (callback != null) {
 			new getAndroidAdvertisingIDInfo(callback).execute(IS_LIMIT_AD_TRACKING_ENABLED);
 		}
 	}
 
 	@Kroll.method
-	public void getAndroidAdId(KrollFunction callback) {
+	public void getAndroidAdId(KrollFunction callback)
+	{
 		if (callback != null) {
 			new getAndroidAdvertisingIDInfo(callback).execute(ANDROID_ADVERTISING_ID);
 		}
 	}
 
-	private void invokeAIDClientInfoCallback(AdvertisingIdClient.Info aaClientIDInfo, String responseKey, KrollFunction callback) {
+	private void invokeAIDClientInfoCallback(AdvertisingIdClient.Info aaClientIDInfo, String responseKey,
+											 KrollFunction callback)
+	{
 		KrollDict callbackDictionary = new KrollDict();
 		Object responseValue = null;
 		switch (responseKey) {
@@ -167,19 +212,23 @@ public class AdmobModule extends KrollModule
 		callback.callAsync(getKrollObject(), callbackDictionary);
 	}
 
-	private class getAndroidAdvertisingIDInfo extends AsyncTask<String, Integer, String> {
+	private class getAndroidAdvertisingIDInfo extends AsyncTask<String, Integer, String>
+	{
 
 		private AdvertisingIdClient.Info aaClientIDInfo = null;
 		private KrollFunction aaInfoCallback;
 
-		public getAndroidAdvertisingIDInfo(KrollFunction infoCallback) {
+		public getAndroidAdvertisingIDInfo(KrollFunction infoCallback)
+		{
 			this.aaInfoCallback = infoCallback;
 		}
 
 		@Override
-		protected String doInBackground(String... responseKey) {
+		protected String doInBackground(String... responseKey)
+		{
 			try {
-				aaClientIDInfo = AdvertisingIdClient.getAdvertisingIdInfo(TiApplication.getInstance().getApplicationContext());
+				aaClientIDInfo =
+					AdvertisingIdClient.getAdvertisingIdInfo(TiApplication.getInstance().getApplicationContext());
 				return responseKey[0];
 			} catch (IOException | GooglePlayServicesNotAvailableException | GooglePlayServicesRepairableException e) {
 				e.printStackTrace();
@@ -188,7 +237,8 @@ public class AdmobModule extends KrollModule
 		}
 
 		@Override
-		protected void onPostExecute(String responseKey) {
+		protected void onPostExecute(String responseKey)
+		{
 			if (aaClientIDInfo != null) {
 				invokeAIDClientInfoCallback(aaClientIDInfo, responseKey, aaInfoCallback);
 			}
@@ -343,8 +393,7 @@ public class AdmobModule extends KrollModule
 
 	// clang-format off
 	@Kroll.setProperty
-	@Kroll.method
-	public void setTagForUnderAgeOfConsent(boolean underAgeOfConsent)
+	public void setIsTaggedForUnderAgeOfConsent(boolean underAgeOfConsent)
 	// clang-format on
 	{
 		Context appContext = TiApplication.getInstance().getApplicationContext();
@@ -352,7 +401,16 @@ public class AdmobModule extends KrollModule
 	}
 
 	@Kroll.method
-	public boolean isTaggedForUnderAgeOfConsent()
+	public void setTagForUnderAgeOfConsent(boolean underAgeOfConsent)
+	{
+		Log.w(TAG, "setTagForUnderAgeOfConsent is deprecated. Use <setIsTaggedForUnderAgeOfConsent> property instead.");
+		Context appContext = TiApplication.getInstance().getApplicationContext();
+		ConsentInformation.getInstance(appContext).setTagForUnderAgeOfConsent(underAgeOfConsent);
+	}
+
+	@Kroll.getProperty
+	@Kroll.method
+	public boolean getIsTaggedForUnderAgeOfConsent()
 	{
 		Context appContext = TiApplication.getInstance().getApplicationContext();
 		return ConsentInformation.getInstance(appContext).isTaggedForUnderAgeOfConsent();
@@ -411,7 +469,7 @@ public class AdmobModule extends KrollModule
 
 			dict.put("identifier", adProvider.getId());
 			dict.put("name", adProvider.getName());
-			dict.put("privacyPolicyURL", adProvider.getPrivacyPolicyUrl());
+			dict.put("privacyPolicyURL", adProvider.getPrivacyPolicyUrlString());
 
 			result[i] = dict;
 		}
@@ -419,4 +477,120 @@ public class AdmobModule extends KrollModule
 		return result;
 	}
 
+	public static Bundle mapToBundle(Map<String, Object> map)
+	{
+		if (map == null) {
+			return new Bundle();
+		}
+
+		Bundle bundle = new Bundle(map.size());
+
+		for (String key : map.keySet()) {
+			Object val = map.get(key);
+			if (val == null) {
+				bundle.putString(key, null);
+			} else if (val instanceof TiBlob) {
+				bundle.putByteArray(key, ((TiBlob) val).getBytes());
+			} else if (val instanceof TiBaseFile) {
+				try {
+					bundle.putByteArray(key, ((TiBaseFile) val).read().getBytes());
+				} catch (IOException e) {
+					Log.e(TAG, "Unable to put '" + key + "' value into bundle: " + e.getLocalizedMessage(), e);
+				}
+			} else if (val instanceof AdmobSizeProxy) {
+				//
+			} else {
+				bundle.putString(key, TiConvert.toString(val));
+			}
+		}
+
+		return bundle;
+	}
+
+	public static AdRequest.Builder createRequestBuilderWithOptions(KrollDict options)
+	{
+		AdRequest.Builder adRequestBuilder = new AdRequest.Builder();
+		if (options == null) {
+			return adRequestBuilder;
+		}
+		Bundle bundle = createAdRequestExtrasBundleFromDictionary(options.getKrollDict("extras"));
+		if (bundle.size() > 0) {
+			adRequestBuilder.addNetworkExtrasBundle(AdMobAdapter.class, bundle);
+		}
+		// Handle keywords
+		if (options.containsKeyAndNotNull("keywords")) {
+			String[] keywords = options.getStringArray("keywords");
+			for (int i = 0; i < keywords.length; i++) {
+				adRequestBuilder.addKeyword(keywords[i]);
+			}
+		}
+		// Handle contentURL
+		if (options.containsKeyAndNotNull("contentURL")) {
+			adRequestBuilder.setContentUrl(options.getString("contentURL"));
+		}
+		// Handle tagForChildDirectedTreatment
+		if (options.containsKeyAndNotNull("tagForChildDirectedTreatment")) {
+			adRequestBuilder.tagForChildDirectedTreatment(options.getBoolean("tagForChildDirectedTreatment"));
+		}
+		// Handle requestAgent
+		if (options.containsKeyAndNotNull("requestAgent")) {
+			adRequestBuilder.setRequestAgent(options.getString("requestAgent"));
+		}
+		// Handle testDevices
+		if (options.containsKeyAndNotNull("testDevices")) {
+			String[] testDevices = options.getStringArray("testDevices");
+			for (int i = 0; i < testDevices.length; i++) {
+				adRequestBuilder.addTestDevice(testDevices[i]);
+			}
+		}
+		return adRequestBuilder;
+	}
+
+	// create the adRequest extras bundle
+	private static Bundle createAdRequestExtrasBundleFromDictionary(KrollDict extrasDictionary)
+	{
+		Bundle bundle = new Bundle();
+		if (extrasDictionary.containsKey(AdmobModule.PROPERTY_COLOR_BG)) {
+			bundle.putString("color_bg", convertColorProp(extrasDictionary.getString(AdmobModule.PROPERTY_COLOR_BG)));
+		}
+		if (extrasDictionary.containsKey(AdmobModule.PROPERTY_COLOR_BG_TOP)) {
+			bundle.putString("color_bg_top",
+							 convertColorProp(extrasDictionary.getString(AdmobModule.PROPERTY_COLOR_BG_TOP)));
+		}
+		if (extrasDictionary.containsKey(AdmobModule.PROPERTY_COLOR_BORDER)) {
+			bundle.putString("color_border",
+							 convertColorProp(extrasDictionary.getString(AdmobModule.PROPERTY_COLOR_BORDER)));
+		}
+		if (extrasDictionary.containsKey(AdmobModule.PROPERTY_COLOR_TEXT)) {
+			bundle.putString("color_text",
+							 convertColorProp(extrasDictionary.getString(AdmobModule.PROPERTY_COLOR_TEXT)));
+		}
+		if (extrasDictionary.containsKey(AdmobModule.PROPERTY_COLOR_LINK)) {
+			bundle.putString("color_link",
+							 convertColorProp(extrasDictionary.getString(AdmobModule.PROPERTY_COLOR_LINK)));
+		}
+		if (extrasDictionary.containsKey(AdmobModule.PROPERTY_COLOR_URL)) {
+			bundle.putString("color_url", convertColorProp(extrasDictionary.getString(AdmobModule.PROPERTY_COLOR_URL)));
+		}
+		return bundle;
+	}
+
+	// modifies the color prop -- removes # and changes constants into hex values
+	public static String convertColorProp(String color)
+	{
+		color = color.replace("#", "");
+		if (color.equals("white"))
+			color = "FFFFFF";
+		if (color.equals("red"))
+			color = "FF0000";
+		if (color.equals("blue"))
+			color = "0000FF";
+		if (color.equals("green"))
+			color = "008000";
+		if (color.equals("yellow"))
+			color = "FFFF00";
+		if (color.equals("black"))
+			color = "000000";
+		return color;
+	}
 }