+++ android/src/ti/admob/ViewProxy.java
@@ -7,82 +7,22 @@
  */
 package ti.admob;
 
-import org.appcelerator.kroll.KrollDict;
 import org.appcelerator.kroll.annotations.Kroll;
-import org.appcelerator.titanium.TiLifecycle.OnLifecycleEvent;
-import org.appcelerator.titanium.proxy.TiViewProxy;
-import org.appcelerator.titanium.TiBaseActivity;
 import org.appcelerator.kroll.common.Log;
-import org.appcelerator.titanium.view.TiUIView;
 
-import android.app.Activity;
-
+/**
+ * This class is left for backwards compatibility until
+ * the deprecated createView() method for BannerView ad
+ * is removed from the module. Users should use createBannerView()
+*/
 @Kroll.proxy(creatableInModule = AdmobModule.class)
-public class ViewProxy extends TiViewProxy implements OnLifecycleEvent
+public class ViewProxy extends BannerViewProxy
 {
-	private View adMob;
-	private static final String TAG = "AdMobViewProxy";
+
+	protected static final String TAG = "AdMobViewProxy";
 
 	public ViewProxy()
 	{
-		super();
-	}
-
-	@Override
-	protected KrollDict getLangConversionTable()
-	{
-		KrollDict table = new KrollDict();
-		table.put("title", "titleid");
-		return table;
-	}
-
-	@Override
-	public TiUIView createView(Activity activity)
-	{
-		adMob = new View(this);
-		((TiBaseActivity) activity).addOnLifecycleEventListener(this);
-		return adMob;
-	}
-
-	@Kroll.method
-	public void requestAd(@Kroll.argument(optional = true) KrollDict parameters)
-	{
-		Log.d(TAG, "requestAd()");
-		adMob.requestAd(parameters);
-	}
-
-	@Kroll.method
-	public void requestTestAd()
-	{
-		Log.d(TAG, "requestTestAd(): ");
-		adMob.requestTestAd();
-	}
-
-	@Override
-	public void onDestroy(Activity activity)
-	{
-		adMob.destroy();
-	}
-
-	@Override
-	public void onPause(Activity activity)
-	{
-		adMob.pause();
-	}
-
-	@Override
-	public void onResume(Activity activity)
-	{
-		adMob.resume();
-	}
-
-	@Override
-	public void onStart(Activity activity)
-	{
-	}
-
-	@Override
-	public void onStop(Activity activity)
-	{
+		Log.w(TAG, "ViewProxy has been deprecated. Use createBannerView instead.");
 	}
 }