+++ ios/platform/GoogleMobileAds.framework/Headers/GADAdReward.h
@@ -7,7 +7,10 @@
 
 #import <Foundation/Foundation.h>
 
-/// Reward information for GADRewardBasedVideoAd ads.
+/// A block to be executed when the user earns a reward.
+typedef void (^GADUserDidEarnRewardHandler)(void);
+
+/// Ad reward information.
 @interface GADAdReward : NSObject
 
 /// Type of the reward.