+++ ios/platform/GoogleMobileAds.framework/Headers/GADFullScreenContentDelegate.h
@@ -27,6 +27,9 @@
 
 @optional
 
+/// Tells the delegate that an impression has been recorded for the ad.
+- (void)adDidRecordImpression:(nonnull id<GADFullScreenPresentingAd>)ad;
+
 /// Tells the delegate that the ad failed to present full screen content.
 - (void)ad:(nonnull id<GADFullScreenPresentingAd>)ad
     didFailToPresentFullScreenContentWithError:(nonnull NSError *)error;