+++ ios/platform/GoogleMobileAds.framework/Headers/GADInterstitial.h
@@ -57,7 +57,7 @@
 /// ad request succeeds or fails.
 @property(nonatomic, readonly, nullable) GADResponseInfo *responseInfo;
 
-/// Called when the ad is estimated to have earned money. Available for whitelisted accounts only.
+/// Called when the ad is estimated to have earned money. Available for allowlisted accounts only.
 @property(nonatomic, nullable, copy) GADPaidEventHandler paidEventHandler;
 
 /// Presents the interstitial ad which takes over the entire screen until the user dismisses it.