+++ ios/platform/GoogleMobileAds.framework/Headers/GADUnifiedNativeAd.h
@@ -68,7 +68,7 @@
 /// Information about the ad response that returned the ad.
 @property(nonatomic, readonly, nonnull) GADResponseInfo *responseInfo;
 
-/// Called when the ad is estimated to have earned money. Available for whitelisted accounts only.
+/// Called when the ad is estimated to have earned money. Available for allowlisted accounts only.
 @property(nonatomic, nullable, copy) GADPaidEventHandler paidEventHandler;
 
 /// Indicates whether custom Mute This Ad is available for the native ad.